+++ app/src/main/java/ps/blob/blobsps/Blob/EnemyBlob.java
@@ -8,17 +8,17 @@ import ps.blob.blobsps.Special.Special;
 public class EnemyBlob extends Blob {
     protected String blobName;
     protected final static int BLOBTYPE = 2;
-    protected int hp;
+    protected double hp;
     protected int sp;
-    protected int atk;
-    protected int def;
+    protected double atk;
+    protected double def;
     protected Special special;
     protected double rarity;
     protected boolean recipe;
     protected int tier;
     protected String imageReference;
 
-    public EnemyBlob(String name, int hp, int sp, int atk, int def, double rarity, boolean recipe, String image, int special) {
+    public EnemyBlob(String name, double hp, int sp, double atk, double def, double rarity, boolean recipe, String image, int special) {
         super(name, hp, sp, atk, def, rarity, false, image, null, special);
     }
