+++ app/src/main/java/ps/blob/blobsps/Blob/PersonalBlob.java
@@ -9,17 +9,17 @@ import ps.blob.blobsps.Special.Special;
 public class PersonalBlob extends Blob {
     protected String blobName;
     protected final static int BLOBTYPE = 1;
-    protected int hp;
+    protected double hp;
     protected int sp;
-    protected int atk;
-    protected int def;
+    protected double atk;
+    protected double def;
     protected Special special;
     protected double rarity;
     protected boolean recipe;
     protected int tier;
     protected String imageReference;
 
-    public PersonalBlob(String name, int hp, int sp, int atk, int def, double rarity, boolean recipe, String image, int special) {
+    public PersonalBlob(String name, double hp, int sp, double atk, double def, double rarity, boolean recipe, String image, int special) {
         super(name, hp, sp, atk, def, rarity, false, image, null, special);
     }
